+++ app/policies/project_policy.rb
@@ -1,22 +1,6 @@
-class ProjectPolicy < ApplicationPolicy
+class ProjectPolicy < ActivityPolicy
   def index?
     true
-  end
-
-  def show?
-    true
-  end
-
-  def create?
-    partner_organisation_user?
-  end
-
-  def update?
-    partner_organisation_user?
-  end
-
-  def destroy?
-    false
   end
 
   def redact_from_iati?