+++ app/policies/third_party_project_policy.rb
@@ -1,11 +1,2 @@
 class ThirdPartyProjectPolicy < ProjectPolicy
-  def create?
-    partner_organisation_user? && editable_report?
-  end
-
-  private
-
-  def editable_report?
-    Report.editable.exists?(organisation: record.organisation, fund: record.associated_fund)
-  end
 end